backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
psycopg2-binary>=2.9.10
pydantic>=2.10.0
httpx>=0.28.0
click>=8.1.7
python-json-logger>=3.0.0
requests>=2.32.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
